Place extensions in separate file
